[app.js]编写了profilemodel的数据，重新build了app.js，对后台的ID改为assetsId

+++ ssh-extjs/src/main/java/com/ssh/demo/staff/dao/IAssetsDao.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 import com.ssh.demo.staff.entity.Assets;
 
 @Component
-public interface IAssetsDao extends PagingAndSortingRepository<Assets, Long>,JpaSpecificationExecutor<Assets> {
+public interface IAssetsDao extends PagingAndSortingRepository<Assets, Integer>,JpaSpecificationExecutor<Assets> {
 //
 //	public Page<Assets> findByAssetsNumberLikeAndPrice(String assetsNumber,Double assetsPrice,Pageable pageable);
 //
